refactor: consolidate neo4j schema into a single place
